requirements.txt
Flask==1.1.1
gunicorn==19.9.0
Jinja2==2.10.3
Werkzeug==0.15.5
pytesseract==0.3.0
numpy>=1.9.2
Pillow==4.3.0
olefile==0.46
click==7.0
itsdangerous==1.1.0
Werkzeug==0.16.0
MarkupSafe==1.1.1
opencv-python==3.4.3